update: make destination ttoken/amount inputs editable
